+++ packages/avatar/stories/avatar.stories.tsx
@@ -1,7 +1,6 @@
-import * as React from 'react';
 import { Stack } from '@nature-ui/layout';
-
-import { Ai } from '@nature-ui/icons';
+import * as React from 'react';
+import { AiOutlineUser } from 'react-icons/ai';
 import { Avatar, AvatarBadge, AvatarGroup } from '../src';
 
 export default {
@@ -110,11 +109,11 @@ export const avatarGroup = () => (
 export const CustomFallbackAvatar = () => (
   <>
     <AvatarGroup spacing='1rem'>
-      <Avatar color='red-500' icon={<Ai.AiOutlineUser size='20px' />} />
+      <Avatar color='red-500' icon={<AiOutlineUser size='20px' />} />
       <Avatar color='teal-500' className='mb-6' />
     </AvatarGroup>
     <AvatarGroup>
-      <Avatar color='red-500' icon={<Ai.AiOutlineUser />} />
+      <Avatar color='red-500' icon={<AiOutlineUser />} />
       <Avatar color='teal-500' />
     </AvatarGroup>
   </>