+++ packages/button/src/button.tsx
@@ -119,10 +119,14 @@ export const ButtonSpinner = (
 ) => {
   const { className = '', label, spinner, ...rest } = props;
 
-  const _className = clsx(className, 'align-middle', {
-    absolute: !label,
-    relative: label,
-  });
+  const _className = clsx(
+    'align-middle',
+    {
+      absolute: !label,
+      relative: label,
+    },
+    className,
+  );
 
   return (
     <span className={_className} {...rest}>
@@ -201,16 +205,24 @@ export const Button = React.forwardRef(
     let BTNClass: string;
 
     if (variant === 'none') {
-      BTNClass = clsx(className, {
-        [STYLES.disabled]: isDisabled || isLoading,
-      });
+      BTNClass = clsx(
+        {
+          [STYLES.disabled]: isDisabled || isLoading,
+        },
+        className,
+      );
     } else {
-      BTNClass = clsx(className, rippleEffect, DEFAULT_CLASS, {
-        [_sizes]: !_link,
-        [STYLES[variant]]: variant,
-        [STYLES.disabled]: isDisabled || isLoading,
-        [_padding]: !isIconButton && !_link,
-      });
+      BTNClass = clsx(
+        rippleEffect,
+        DEFAULT_CLASS,
+        {
+          [_sizes]: !_link,
+          [STYLES[variant]]: variant,
+          [STYLES.disabled]: isDisabled || isLoading,
+          [_padding]: !isIconButton && !_link,
+        },
+        className,
+      );
     }
 
     const defaults = {