+++ packages/layout/src/stack.tsx
@@ -1,3 +1,4 @@
+/** ** */
 import { clsx, forwardRef, nature, PropsOf } from '@nature-ui/system';
 import { getValidChildren, StringOrNumber, __DEV__ } from '@nature-ui/utils';
 import * as React from 'react';
@@ -63,11 +64,14 @@ export const Stack = forwardRef<StackProps>((props, ref) => {
     return child;
   });
 
-  const _className = clsx(className, {
-    flex: !responsive,
-    'block sm:flex': responsive,
-    [`flex-${direction}`]: direction && !responsive,
-  });
+  const _className = clsx(
+    {
+      flex: !responsive,
+      'block sm:flex': responsive,
+      [`flex-${direction}`]: direction && !responsive,
+    },
+    className,
+  );
 
   return (
     <DivTag className={_className} {...rest} ref={ref}>