+++ packages/alert-dialog/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @nature-ui/alert-dialog
 
+## 1.1.2
+
+### Patch Changes
+
+- 0922ed4: - Proper className placement
+  - Input: added `focusBorderColor`, and `focusErrorColor` feature
+- Updated dependencies [0922ed4]
+  - @nature-ui/modal@1.1.3
+
 ## 1.1.1
 
 ### Patch Changes