+++ packages/container/CHANGELOG.md
@@ -1,11 +1,13 @@
 # @nature-ui/container
 
-## 1.1.2
+## 1.1.3
 
 ### Patch Changes
 
-- Updated dependencies [undefined]
-  - @nature-ui/system@2.0.0
+- 0922ed4: - Proper className placement
+  - Input: added `focusBorderColor`, and `focusErrorColor` feature
+- Updated dependencies [0922ed4]
+  - @nature-ui/system@2.0.1
 
 ## 1.1.1