+++ packages/drawer/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @nature-ui/drawer
 
+## 1.1.1
+
+### Patch Changes
+
+- 0922ed4: - Proper className placement
+  - Input: added `focusBorderColor`, and `focusErrorColor` feature
+- Updated dependencies [0922ed4]
+  - @nature-ui/modal@1.1.3
+  - @nature-ui/transition@1.1.1
+
 ## 1.1.0
 
 ### Minor Changes