+++ packages/image/CHANGELOG.md
@@ -1,11 +1,14 @@
 # @nature-ui/image
 
-## 1.2.2
+## 1.2.3
 
 ### Patch Changes
 
-- Updated dependencies [undefined]
-  - @nature-ui/system@2.0.0
+- 0922ed4: - Proper className placement
+  - Input: added `focusBorderColor`, and `focusErrorColor` feature
+- Updated dependencies [0922ed4]
+  - @nature-ui/hooks@1.1.1
+  - @nature-ui/system@2.0.1
 
 ## 1.2.1